requirements.txt
tflite-runtime==2.5.0
flask==3.0.0
flask-cors==4.0.0
pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3